+++ middleware/cookie-message.js
@@ -1,97 +1,42 @@
-const cookieParser = require('cookie-parser');
-const { setCookie, clearAllCookies } = require('../utils/cookies');
-const logger = require('@dwp/govuk-casa/lib/Logger')('ds1500');
-const { COOKIE_UPDATE_QUERY_PARAM } = require('../lib/constants')
-// Load app config from `.env`
+const qs = require('querystring');
 
-// eslint-disable-next-line max-len
-module.exports = (app, proxyUrl, proxyMountUrl, consentCookieName, cookiePolicy, cookieConsent, useTLS) => {
+module.exports = (app, consentCookieName, waypoints, mountUrl = '/', proxyMountUrl = mountUrl) => {
+  const reProxyMountUrl = new RegExp(`^${proxyMountUrl}`);
+  const sanitiseUrl = (url) => url.replace(reProxyMountUrl, mountUrl).replace(/\/+/g, '/');
+
   // URL to cookie policy page
-  const cookieDetailsUrl = `${proxyUrl}cookie-details`;
-
-  // ignore in test mode
-  if (process.env.NODE_ENV !== 'test') {
-    app.use(cookieParser());
-  }
+  const cookiePolicyUrl = `${mountUrl}${waypoints.COOKIE_POLICY}`;
 
   // Set template options for cookie consent banner
   app.use((req, res, next) => {
-    const staticPatt = /^\/(assets|css|images|js)/gi;
-    const cookiePolicyUrl = `${proxyUrl}${cookiePolicy}`;
-    if (staticPatt.test(req.path)) {
-      logger.info('Skipping static path');
-      next();
-      return;
-    }
-    const defaults = {
-      COOKIE_CONSENT_CHOICE: null
-    };
-    const cookies = Object.assign({}, defaults, req.cookies);
-    const cookieValue = cookies[consentCookieName]
-    const isChoiceMade = () => {
-      const validChoices = ['accept', 'reject']
-      return cookieValue && validChoices.indexOf(cookieValue) > -1
-    };
     // Get cookie banner flash messages (did you accept / reject)
     if (req.session) {
-      res.locals.cookieChoiceMade = isChoiceMade()
-      req.session.cookieChoiceMade = isChoiceMade();
+      res.locals.cookieChoiceMade = req.session.cookieChoiceMade;
+      req.session.cookieChoiceMade = undefined;
     }
 
     // Add current consent cookie value to templates
     if (req.cookies[consentCookieName]) {
-      res.locals.cookieMessage = cookieValue;
-    } else {
-      res.locals.cookieMessage = 'unset';
+      res.locals.consentCookieValue = req.cookies[consentCookieName];
     }
 
     // Url to submit consent to (used in banner)
-    res.locals.cookieConsentSubmit = cookieConsent;
-    res.locals.showCookieConstentMessage = !isChoiceMade();
-    res.locals.showCookieChoiceConfirmationBanner = parseInt(req.query[COOKIE_UPDATE_QUERY_PARAM], 10) === 1;
-    res.locals.cookiesConsented = cookieValue
-    res.locals.currentPage = req.path;
-    res.locals.previousPage = req.query.previousPage;
-    res.locals.cookiePolicyUrl = cookiePolicyUrl;
+    res.locals.cookieConsentSubmit = waypoints.COOKIE_CONSENT;
+    res.locals.consentCookieName = consentCookieName;
 
     // Set backto query
-    res.locals.cookieDetailsUrl = cookieDetailsUrl
-    res.locals.currentPage = encodeURIComponent(req.path)
-    res.locals.previousPage = req.query.previousPage
-    // Set referrer policy
-    res.set('Referrer-Policy', 'same-origin');
-    next();
-  });
+    const { pathname, search } = new URL(String(req.url), 'https://dummy.test/');
+    const sanitisedPath = sanitiseUrl(pathname);
+    const currentUrl = sanitisedPath + search;
+    res.locals.currentUrl = currentUrl;
 
-  // Handle setting consent cookie from banner submisson
-  app.post(`${proxyMountUrl}${cookieConsent}/:cookieMethod`, (req, res) => {
-    const url = require('url');
-    const cookiePolicyUrl = `${proxyUrl}${cookiePolicy}`;
-    const { cookieMethod } = req.params;
-    let redirectPath
-    const referrer = req.get('Referrer');
-
-    if (referrer) {
-      const { pathname } = new url.URL(referrer)
-      redirectPath = pathname
+    // If already on cookie policy page, don't need set backto again
+    if (sanitisedPath === cookiePolicyUrl) {
+      res.locals.cookiePolicyUrl = currentUrl;
     } else {
-      redirectPath = cookiePolicyUrl
+      res.locals.cookiePolicyUrl = `${cookiePolicyUrl}?${qs.stringify({ backto: currentUrl })}`;
     }
 
-    if (cookieMethod === 'reject' || cookieMethod === 'accept') {
-      setCookie(req, res, consentCookieName, cookieMethod, useTLS);
-      if (cookieConsent === 'reject') {
-        clearAllCookies(req, res)
-      }
-    }
-
-    const qs = {};
-    qs[COOKIE_UPDATE_QUERY_PARAM] = 1;
-    const redirectUrl = url.format({
-      pathname: redirectPath,
-      query: Object.assign({}, qs)
-    });
-
-    res.redirect(redirectUrl);
+    next();
   });
 };